refactor: Moves layout files to layout folder
