+++ apps/web/src/app/layout.tsx
@@ -1,18 +1,18 @@
-import type { Metadata } from 'next';
-import { Rubik } from 'next/font/google';
-import './globals.css';
-import Header from '@/components/header';
-import { cn } from '@/lib/utils';
-import { Providers } from '@/providers';
+import Header from "@/components/layout/header";
+import { cn } from "@/lib/utils";
+import { Providers } from "@/providers";
+import type { Metadata } from "next";
+import { Rubik } from "next/font/google";
+import "./globals.css";
 
 const rubik = Rubik({
-  weight: ['400', '500', '700'],
-  subsets: ['latin'],
+  weight: ["400", "500", "700"],
+  subsets: ["latin"],
 });
 
 export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
+  title: "Create Next App",
+  description: "Generated by create next app",
 };
 
 export default function RootLayout({
@@ -22,7 +22,7 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
-      <body className={cn(rubik.className, 'bg-background text-foreground')}>
+      <body className={cn(rubik.className, "bg-background text-foreground")}>
         <Providers>
           <Header />
           {children}
