+++ apps/web/src/components/layout/header.tsx
@@ -1,12 +1,11 @@
-'use client';
+"use client";
 
-import { useThemeStore } from '@/store/useThemeStore';
-import { Moon } from 'lucide-react';
-import Link from 'next/link';
-import React from 'react';
-import { ConnectWalletButton } from './connect-wallet-button';
-import NavLink from './nav-link';
-import { Button } from './ui/button';
+import { useThemeStore } from "@/store/useThemeStore";
+import { Moon } from "lucide-react";
+import Link from "next/link";
+import { Button } from "../ui/button";
+import { ConnectWalletButton } from "./connect-wallet-button";
+import NavLink from "./nav-link";
 
 const Header = () => {
   const { toggleTheme } = useThemeStore();
@@ -14,7 +13,9 @@ const Header = () => {
   return (
     <header className="flex relative h-16 items-center justify-between bg-content mx-5 mt-6 px-5 rounded-3xl ">
       <Link href="/">
-        <h1 className="text-primary font-bold text-lg font-roboto-mono text-center">Bleu</h1>
+        <h1 className="text-primary font-bold text-lg font-roboto-mono text-center">
+          Bleu
+        </h1>
       </Link>
       <div className="absolute left-1/2 top-1/2 -translate-x-1/2 -translate-y-1/2 flex gap-4 text-lg">
         <NavLink href="/">Home</NavLink>
